seperating avatar name parsing
