+++ js/viewerInfo.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var os = require('os');
 
 var packageJSON = require('../package.json');
